Cleanup/Restore bot button messages

+++ ui/ui_util.py
@@ -2,7 +2,7 @@
 import asyncio
 from datetime import datetime, date
 import logging
-from nextcord.ext import commands
+import math
 import nextcord
 import re
 import validators
@@ -13,85 +13,24 @@ from db.zBot_db_orm import *
 # UTILITY TYPES
 ########################################################################################################################
 class zField():
-    def __init__(self, name: str, label: str, default_value, required: bool):
-        self.name = name
+    def __init__(self, custom_id: str, label: str, default_value, required: bool, placeholder=None):
+        self.custom_id = custom_id
         self.label = label
         self.default_value = default_value
         self.required = required
+        self.placeholder = placeholder
 
 FieldList = list[zField]
 FieldDict = dict[str, nextcord.ui.TextInput]
 ModalDataList = dict[str, str]
 SelectList = list[nextcord.SelectOption]
 
-VarTypeInt      = 1
-VarTypeStr      = 2
-VarTypeGameTime = 3
-VarTypeDateTime = 4
-
-VartypeStrDict = {
-    VarTypeInt:       "Integer",
-    VarTypeStr:       "String",
-    VarTypeGameTime:  "Game Time (H:MM:SS)",
-    VarTypeDateTime:  "Date/Time",
-}
-
-VartypeSelectOptionList = [
-    nextcord.SelectOption(label="Integer", value=VarTypeInt, description="Integer Value"),
-    nextcord.SelectOption(label="String", value=VarTypeStr, description="String text up to 255 characters in length"),
-    nextcord.SelectOption(label="Game Time (H:MM:SS)", value=VarTypeGameTime, description="Game time expressed in 'H:MM:SS' format"),
-    nextcord.SelectOption(label="Date/Time", value=VarTypeDateTime, description="Date & Time string, typically something like 'YYYY-MM-DD HH:MM:SS'"),
-]
-
 ########################################################################################################################
 # UTILITY FUNCTIONS
 ########################################################################################################################
-def get_category_select_list(server_id):
-    # Get the list of categories for this server
-    categories = AsyncRaceCategory.select().where(AsyncRaceCategory.server_id == server_id)
-
-    # Populate the SelectOption list with the category information
-    select_list = []
-    for c in categories:
-        select_list.append(nextcord.SelectOption(label=c.name, value=c.id, description=c.description))
-    return select_list
-
-#####################################################################################################################
-def get_race_select_list(server_id):
-    # Get the list of race for this server
-    races = AsyncRace.select().where(AsyncRace.server_id == server_id)
-
-    # Populate the SelectOption list with the race information
-    select_list = []
-    for r in races:
-        select_list.append(nextcord.SelectOption(label=f"{r.id} - {r.description[:15]}", value=r.id, description=r.description))
-    return select_list
-
-#####################################################################################################################
-def get_extra_info_type_select_list(server_id, race =None):
-    # Get the list of extra info types for this server
-    types = AsyncRaceExtraInfoType.select().where(AsyncRaceExtraInfoType.server_id == server_id)
-    default_list = []
-    if race is not None:
-        for r in race.extra_info_assignments:
-            default_list.append(r.info_type_id.id)
-
-    # Populate the SelectOption list with the race information
-    select_list = []
-    for t in types:
-        select_list.append(nextcord.SelectOption(label=f"{t.name}", value=t.id, description=t.description, default=t.id in default_list))
-    return select_list
-
-#####################################################################################################################
-def get_race_extra_info(race_id):
-    infos = []
-    if race_id is not None:
-        infos = AsyncRaceExtraInfoAssignment.select().where(AsyncRaceExtraInfoAssignment.race_id == race_id)
-    return infos
-
-####################################################################################################################
-# Takes an AsyncRaceMessage, finds the corresponding Discord message and deletes both the message and DB entry
-async def delete_message(server, async_race_msg_id):
+# Takes an AsyncRaceMessage, finds the corresponding Discord message and deletes message and optionally deletes or 
+# zeroes the DB entry
+async def delete_message(server, async_race_msg_id, zero_db = False):
     async_race_msg = None
     if async_race_msg_id is not None:
         try:
@@ -107,7 +46,11 @@ async def delete_message(server, async_race_msg_id):
                 await msg.delete()
             except:
                 logging.info(f"Failed to find message with message ID {async_race_msg.message_id}")
-        async_race_msg.delete_instance()
+        if zero_db:
+            async_race_msg.message_id = 0
+            async_race_msg.save()
+        else:
+            async_race_msg.delete_instance()
 
 #####################################################################################################################
 async def has_text_channel_permission(user_id, server, channel):
@@ -137,60 +80,65 @@ def get_role_select_list(server):
         select_list.append(nextcord.SelectOption(label=r.name, value=r.id, description=r.name))
     return select_list
 
-#####################################################################################################################
-def get_race(race_id):
-    race = None
-    if race_id is not None:
-        try:
-            race = AsyncRace.select().where(AsyncRace.id == race_id).get()
-        except:
-            logging.info(f"Could not find race ID {race_id}")
-    return race
+########################################################################################################################
+# BASE CLASSES
+########################################################################################################################
+# This Modal (form) will display a form that has fields matching the provided item list. On completion it will call
+# the provided submit_handler function, passing a pointer to itself and the interaction object.
+class zModal(nextcord.ui.Modal):
+    def __init__(self, fields: FieldDict, submit_handler, title: str):
+        super().__init__(title, timeout=None)
+        self.fields = fields
+        self.submit_handler = submit_handler
+
+        for v in self.fields.values():
+            self.add_item(v)
+
+    async def callback(self, interaction: nextcord.Interaction) -> None:
+        await self.submit_handler(interaction, self)
 
 #####################################################################################################################
-def get_category(category_id):
-    category = None
-    if category_id is not None:
-        try:
-            category = AsyncRaceCategory.select().where(AsyncRaceCategory.id == category_id).get()
-        except:
-            logging.info(f"Could not find category ID {category_id}")
-    return category
+# This Select (drop down selection) will display a drop down with the string options provided. This variant will
+# only allow the user to select a single option from the list. On completion it will call the provided
+# submit_handler function, passing the value for the option chosen and the interaction object.
+class zSingleSelect(nextcord.ui.Select):
+    def __init__(self, select_list: SelectList, submit_handler, placeholder):
+        super().__init__(min_values=1, max_values=1, options=select_list, placeholder=placeholder)
+        self.submit_handler = submit_handler
+
+    async def callback(self, interaction: nextcord.Interaction) -> None:
+        await self.submit_handler(int(interaction.data['values'][0]), interaction)
 
 #####################################################################################################################
-def get_extra_info_type(info_type_id):
-    info_type = None
-    if info_type_id is not None:
-        try:
-            info_type = AsyncRaceExtraInfoType.select().where(AsyncRaceExtraInfoType.id == info_type_id).get()
-        except:
-            logging.info(f"Could not find Info Type ID {info_type_id}")
-    return info_type
+# View which contains a zSingleSelect
+class zSingleSelectView(nextcord.ui.View):
+    def __init__(self, select_list: SelectList, submit_handler, placeholder = None):
+        super().__init__(timeout=None)
+        self.category_select = zSingleSelect(select_list, submit_handler, placeholder)
+        self.add_item(self.category_select)
 
 #####################################################################################################################
-# Returns the current date/time in the format used in the database
-def zBot_now():
-    return datetime.now().isoformat(timespec='minutes').replace('T', ' ')
+# This Select (drop down selection) will display a drop down with the string options provided. This variant will
+# allow the user to select up to `max_values` from the list. On completion it will call the provided submit_handler
+# function, passing a list of the values chosen and the interaction object.
+class zMultiSelect(nextcord.ui.Select):
+    def __init__(self, select_list: SelectList, max_values: int, submit_handler, placeholder):
+        super().__init__(min_values=1, max_values=max_values, options=select_list, placeholder=placeholder)
+        self.submit_handler = submit_handler
+
+    async def callback(self, interaction: nextcord.Interaction) -> None:
+        value_list = []
+        for v in interaction.data['values']:
+            value_list.append(int(v))
+        await self.submit_handler(value_list, interaction)
 
 #####################################################################################################################
-def check_server_assignment_exists(info_type_id, server_id):
-    try:
-        a = AsyncRaceExtraInfoAssignment.select().where(
-            AsyncRaceExtraInfoAssignment.info_type_id == info_type_id and
-            AsyncRaceExtraInfoAssignment.server_id == server_id).get()
-    except:
-        a = None
-    return a is not None
-
-#####################################################################################################################
-def check_category_assignment_exists(info_type_id, category_id):
-    try:
-        a = AsyncRaceExtraInfoAssignment.select().where(
-                AsyncRaceExtraInfoAssignment.info_type_id == info_type_id and
-                AsyncRaceExtraInfoAssignment.category_id == category_id).get()
-    except:
-        a = None
-    return a is not None
+# View which contains a zMultiSelect
+class zMultiSelectView(nextcord.ui.View):
+    def __init__(self, select_list: SelectList, max_values: int, submit_handler, placeholder = None):
+        super().__init__(timeout=None)
+        self.category_select = zMultiSelect(select_list, max_values, submit_handler, placeholder)
+        self.add_item(self.category_select)
 
 ########################################################################################################################
 # View which contains buttons for continuing or cancelling
@@ -213,21 +161,24 @@ class zContinueCancelButtonView(nextcord.ui.View):
 # modal pages if needed. The final set of data values will be sent to the provided `submit_handler` function
 # as a ModalDataList using the same keys used in the original FieldList.
 class zMultiPageModalSender():
-    def __init__():
+    def __init__(self):
         pass
 
     ####################################################################################################################
-    async def send_modal(interaction, fields: FieldList, submit_handler, title: str):
+    async def send_modal(self, interaction, fields: FieldList, submit_handler, title: str):
         self.fields = fields
         self.submit_handler = submit_handler
         self.title = title
         self.field_idx = 0
-        await self.send_next_modal_page(interaction)
+        self.submit_values = []
+        for i in range(0, len(fields)):
+            self.submit_values.append(None)
+        await self.send_modal_page(interaction)
 
     ####################################################################################################################
-    def get_field_index(self, field_name):
+    def get_field_index(self, custom_id):
         for (i, f) in enumerate(self.fields):
-            if f.name == field_name:
+            if f.custom_id == custom_id:
                 return i
         return None
 
@@ -238,16 +189,18 @@ class zMultiPageModalSender():
         # Fill in the field dictionary for the modal we're about to create and send
         while curr_row < 5 and self.field_idx < len(self.fields):
             f = self.fields[self.field_idx]
-            fieldDict[f.name] = nextcord.ui.TextInput(
+            fieldDict[f.custom_id] = nextcord.ui.TextInput(
                 label=f.label,
                 required=f.required,
-                custom_id=f.name,
+                custom_id=f.custom_id,
                 default_value=f.default_value,
+                placeholder=f.placeholder,
                 row=curr_row)
             curr_row += 1
             self.field_idx += 1
         # Create and send the modal
-        modal = zModal(fieldDict, self.on_page_submit, self.title)
+        page = math.trunc(self.field_idx / 4) + 1
+        modal = zModal(fieldDict, self.on_page_submit, self.title + f" [pg. {page}]")
         await interaction.response.send_modal(modal)
 
     ####################################################################################################################
@@ -267,9 +220,8 @@ class zMultiPageModalSender():
                                    ephemeral=True)
         else:
             # Otherwise we're done, so call the originally provided submit handler
-            self.submit_handler(interaction, self.submit_values)
+            await self.submit_handler(interaction, self.submit_values)
 
     ####################################################################################################################
-    async def cancel_submit(interaction):
-        await interaction.send("Cancelled")
-        self.submit_handler(interaction, None)
+    async def cancel_submit(self, interaction):
+        await self.submit_handler(interaction, None)
